# perf: optimize cloning and minor other changes