--- a/lib/mquery.js
+++ b/lib/mquery.js
@@ -76,7 +76,7 @@ function Query(criteria, options) {
 
 let $withinCmd = '$geoWithin';
 Object.defineProperty(Query, 'use$geoWithin', {
-  get: function() { return $withinCmd == '$geoWithin'; },
+  get: function() { return $withinCmd === '$geoWithin'; },
   set: function(v) {
     if (true === v) {
       // mongodb >= 2.4
@@ -176,7 +176,7 @@ Query.prototype.setOptions = function(options) {
     method = methods[i];
 
     // use methods if exist (safer option manipulation)
-    if ('function' == typeof this[method]) {
+    if ('function' === typeof this[method]) {
       const args = Array.isArray(options[method])
         ? options[method]
         : [options[method]];
@@ -278,7 +278,7 @@ Query.prototype.where = function() {
 
   const type = typeof arguments[0];
 
-  if ('string' == type) {
+  if ('string' === type) {
     this._path = arguments[0];
 
     if (2 === arguments.length) {
@@ -288,7 +288,7 @@ Query.prototype.where = function() {
     return this;
   }
 
-  if ('object' == type && !Array.isArray(arguments[0])) {
+  if ('object' === type && !Array.isArray(arguments[0])) {
     return this.merge(arguments[0]);
   }
 
@@ -675,7 +675,7 @@ Query.prototype.exists = function() {
  */
 
 Query.prototype.elemMatch = function() {
-  if (null == arguments[0])
+  if (null === arguments[0] || undefined === arguments[0])
     throw new TypeError('Invalid argument');
 
   let fn, path, criteria;
@@ -830,7 +830,7 @@ Query.prototype.box = function() {
 Query.prototype.polygon = function() {
   let val, path;
 
-  if ('string' == typeof arguments[0]) {
+  if ('string' === typeof arguments[0]) {
     // polygon('loc', [],[],[])
     val = Array.from(arguments);
     path = val.shift();
@@ -1058,10 +1058,10 @@ Query.prototype.intersects = function intersects() {
  */
 
 Query.prototype.geometry = function geometry() {
-  if (!('$within' == this._geoComparison ||
-        '$geoWithin' == this._geoComparison ||
-        '$near' == this._geoComparison ||
-        '$geoIntersects' == this._geoComparison)) {
+  if (!('$within' === this._geoComparison ||
+        '$geoWithin' === this._geoComparison ||
+        '$near' === this._geoComparison ||
+        '$geoIntersects' === this._geoComparison)) {
     throw new Error('geometry() must come after `within()`, `intersects()`, or `near()');
   }
 
@@ -1112,6 +1112,7 @@ Query.prototype.geometry = function geometry() {
  * @see SchemaType
  * @api public
  */
+const whitespaceRE = /\s+/;
 
 Query.prototype.select = function select() {
   let arg = arguments[0];
@@ -1127,15 +1128,15 @@ Query.prototype.select = function select() {
   const type = typeof arg;
   let i, len;
 
-  if (('string' == type || utils.isArgumentsObject(arg)) &&
-    'number' == typeof arg.length || Array.isArray(arg)) {
-    if ('string' == type)
-      arg = arg.split(/\s+/);
+  if (('string' === type || utils.isArgumentsObject(arg)) &&
+    'number' === typeof arg.length || Array.isArray(arg)) {
+    if ('string' === type)
+      arg = arg.split(whitespaceRE);
 
     for (i = 0, len = arg.length; i < len; ++i) {
       let field = arg[i];
       if (!field) continue;
-      const include = '-' == field[0] ? 0 : 1;
+      const include = '-' === field[0] ? 0 : 1;
       if (include === 0) field = field.substring(1);
       fields[field] = include;
     }
@@ -1247,7 +1248,7 @@ Query.prototype.sort = function(arg) {
   // .sort([['field', 1], ['test', -1]])
   if (Array.isArray(arg)) {
     len = arg.length;
-    for (i = 0; i < arg.length; ++i) {
+    for (i = 0; i < len; ++i) {
       if (!Array.isArray(arg[i])) {
         throw new Error('Invalid sort() argument, must be array of arrays');
       }
@@ -1257,13 +1258,13 @@ Query.prototype.sort = function(arg) {
   }
 
   // .sort('field -test')
-  if (1 === arguments.length && 'string' == type) {
-    arg = arg.split(/\s+/);
+  if (1 === arguments.length && 'string' === type) {
+    arg = arg.split(whitespaceRE);
     len = arg.length;
     for (i = 0; i < len; ++i) {
       field = arg[i];
       if (!field) continue;
-      const ascend = '-' == field[0] ? -1 : 1;
+      const ascend = '-' === field[0] ? -1 : 1;
       if (ascend === -1) field = field.substring(1);
       push(this.options, field, ascend);
     }
@@ -2104,7 +2105,7 @@ Query.prototype.distinct = function(criteria, field, callback) {
     switch (typeof field) {
       case 'function':
         callback = field;
-        if ('string' == typeof criteria) {
+        if ('string' === typeof criteria) {
           field = criteria;
           criteria = undefined;
         }
@@ -2128,7 +2129,7 @@ Query.prototype.distinct = function(criteria, field, callback) {
     }
   }
 
-  if ('string' == typeof field) {
+  if ('string' === typeof field) {
     this._distinct = field;
   }
 
@@ -2241,13 +2242,13 @@ Query.prototype.update = function update(criteria, doc, options, callback) {
 
   switch (arguments.length) {
     case 3:
-      if ('function' == typeof options) {
+      if ('function' === typeof options) {
         callback = options;
         options = undefined;
       }
       break;
     case 2:
-      if ('function' == typeof doc) {
+      if ('function' === typeof doc) {
         callback = doc;
         doc = criteria;
         criteria = undefined;
@@ -2299,13 +2300,13 @@ Query.prototype.updateMany = function updateMany(criteria, doc, options, callbac
 
   switch (arguments.length) {
     case 3:
-      if ('function' == typeof options) {
+      if ('function' === typeof options) {
         callback = options;
         options = undefined;
       }
       break;
     case 2:
-      if ('function' == typeof doc) {
+      if ('function' === typeof doc) {
         callback = doc;
         doc = criteria;
         criteria = undefined;
@@ -2357,13 +2358,13 @@ Query.prototype.updateOne = function updateOne(criteria, doc, options, callback)
 
   switch (arguments.length) {
     case 3:
-      if ('function' == typeof options) {
+      if ('function' === typeof options) {
         callback = options;
         options = undefined;
       }
       break;
     case 2:
-      if ('function' == typeof doc) {
+      if ('function' === typeof doc) {
         callback = doc;
         doc = criteria;
         criteria = undefined;
@@ -2414,13 +2415,13 @@ Query.prototype.replaceOne = function replaceOne(criteria, doc, options, callbac
 
   switch (arguments.length) {
     case 3:
-      if ('function' == typeof options) {
+      if ('function' === typeof options) {
         callback = options;
         options = undefined;
       }
       break;
     case 2:
-      if ('function' == typeof doc) {
+      if ('function' === typeof doc) {
         callback = doc;
         doc = criteria;
         criteria = undefined;
@@ -2700,13 +2701,13 @@ Query.prototype.findOneAndUpdate = function(criteria, doc, options, callback) {
 
   switch (arguments.length) {
     case 3:
-      if ('function' == typeof options) {
+      if ('function' === typeof options) {
         callback = options;
         options = {};
       }
       break;
     case 2:
-      if ('function' == typeof doc) {
+      if ('function' === typeof doc) {
         callback = doc;
         doc = criteria;
         criteria = undefined;
@@ -2714,7 +2715,7 @@ Query.prototype.findOneAndUpdate = function(criteria, doc, options, callback) {
       options = undefined;
       break;
     case 1:
-      if ('function' == typeof criteria) {
+      if ('function' === typeof criteria) {
         callback = criteria;
         criteria = options = doc = undefined;
       } else {
@@ -2774,10 +2775,10 @@ Query.prototype.findOneAndRemove = Query.prototype.findOneAndDelete = function(c
   this.op = 'findOneAndRemove';
   this._validate();
 
-  if ('function' == typeof options) {
+  if ('function' === typeof options) {
     callback = options;
     options = undefined;
-  } else if ('function' == typeof conditions) {
+  } else if ('function' === typeof conditions) {
     callback = conditions;
     conditions = undefined;
   }
@@ -2883,13 +2884,13 @@ Query.prototype.exec = function exec(op, callback) {
 
   assert.ok(this.op, 'Missing query type: (find, update, etc)');
 
-  if ('update' == this.op || 'remove' == this.op) {
+  if ('update' === this.op || 'remove' === this.op) {
     callback || (callback = true);
   }
 
   const _this = this;
 
-  if ('function' == typeof callback) {
+  if ('function' === typeof callback) {
     this[this.op](callback);
   } else {
     return new Query.Promise(function(success, error) {