--- a/lib/utils.js
+++ b/lib/utils.js
@@ -4,7 +4,10 @@
  * Module dependencies.
  */
 
-const specialProperties = ['__proto__', 'constructor', 'prototype'];
+const specialProperties = new Set(['__proto__', 'constructor', 'prototype']);
+const primitiveTypes = new Set(['bigint', 'string', 'number', 'boolean', 'symbol', 'undefined']);
+const objectIdVariants = new Set(['ObjectId', 'ObjectID']);
+const dateVariants = new Set(['Date', 'Function']);
 
 /**
  * Clones objects
@@ -16,44 +19,55 @@ const specialProperties = ['__proto__', 'constructor', 'prototype'];
  */
 
 const clone = exports.clone = function clone(obj, options) {
-  if (obj === undefined || obj === null)
+  if (obj === null || primitiveTypes.has(typeof obj)) {
     return obj;
+  }
 
-  if (Array.isArray(obj))
+  if (Array.isArray(obj)) {
     return exports.cloneArray(obj, options);
+  }
+
+  if (obj instanceof Date) {
+    return new Date(obj);
+  }
+
+  if (obj instanceof RegExp) {
+    return new RegExp(obj.source, obj.flags);
+  }
+
+  if (Buffer.isBuffer(obj)) {
+    return Buffer.from(obj.slice());
+  }
 
   if (obj.constructor) {
-    if (/ObjectI[dD]$/.test(obj.constructor.name)) {
-      return 'function' == typeof obj.clone
-        ? obj.clone()
-        : new obj.constructor(obj.id);
+    const constructorName = obj.constructor.name;
+    if (objectIdVariants.has(constructorName)) {
+      return typeof obj.clone === 'function' && obj.clone() ||
+        new obj.constructor(obj.id);
     }
 
-    if (obj.constructor.name === 'ReadPreference') {
+    if (dateVariants.has(constructorName)) {
+      return new obj.constructor(+obj);
+    }
+
+    if (obj._bsontype === 'Binary' && obj.buffer && obj.value) {
+      return typeof obj.clone === 'function' && obj.clone() ||
+        new obj.constructor(obj.value(true), obj.sub_type);
+    }
+
+    if (constructorName === 'ReadPreference') {
       return new obj.constructor(obj.mode, clone(obj.tags, options));
     }
 
-    if ('Binary' == obj._bsontype && obj.buffer && obj.value) {
-      return 'function' == typeof obj.clone
-        ? obj.clone()
-        : new obj.constructor(obj.value(true), obj.sub_type);
-    }
-
-    if ('Date' === obj.constructor.name || 'Function' === obj.constructor.name)
-      return new obj.constructor(+obj);
-
-    if ('RegExp' === obj.constructor.name)
-      return new RegExp(obj);
-
-    if ('Buffer' === obj.constructor.name)
-      return Buffer.from(obj);
   }
 
-  if (isObject(obj))
+  if (isObject(obj)) {
     return exports.cloneObject(obj, options);
+  }
 
-  if (obj.valueOf)
+  if (obj.valueOf) {
     return obj.valueOf();
+  }
 };
 
 /*!
@@ -74,13 +88,13 @@ exports.cloneObject = function cloneObject(obj, options) {
     k = keys[i];
     // Not technically prototype pollution because this wouldn't merge properties
     // onto `Object.prototype`, but avoid properties like __proto__ as a precaution.
-    if (specialProperties.indexOf(k) !== -1) {
+    if (specialProperties.has(k)) {
       continue;
     }
 
     val = clone(obj[k], options);
 
-    if (!minimize || ('undefined' !== typeof val)) {
+    if (!minimize || (typeof val !== 'undefined')) {
       hasKeys || (hasKeys = true);
       ret[k] = val;
     }
@@ -92,11 +106,12 @@ exports.cloneObject = function cloneObject(obj, options) {
 };
 
 exports.cloneArray = function cloneArray(arr, options) {
-  const ret = [],
-      l = arr.length;
+  const len = arr.length,
+      ret = new Array(len);
   let i = 0;
-  for (; i < l; i++)
-    ret.push(clone(arr[i], options));
+  for (; i < len; ++i) {
+    ret[i] = clone(arr[i], options);
+  }
   return ret;
 };
 
@@ -141,10 +156,10 @@ exports.merge = function merge(to, from) {
   const keys = Object.keys(from);
 
   for (const key of keys) {
-    if (specialProperties.indexOf(key) !== -1) {
+    if (specialProperties.has(key)) {
       continue;
     }
-    if ('undefined' === typeof to[key]) {
+    if (typeof to[key] === 'undefined') {
       to[key] = from[key];
     } else {
       if (exports.isObject(from[key])) {
@@ -168,10 +183,10 @@ exports.mergeClone = function mergeClone(to, from) {
   const keys = Object.keys(from);
 
   for (const key of keys) {
-    if (specialProperties.indexOf(key) !== -1) {
+    if (specialProperties.has(key)) {
       continue;
     }
-    if ('undefined' === typeof to[key]) {
+    if (typeof to[key] === 'undefined') {
       to[key] = clone(from[key]);
     } else {
       if (exports.isObject(from[key])) {
@@ -197,26 +212,16 @@ exports.mergeClone = function mergeClone(to, from) {
  * @param {String} pref
  */
 
-exports.readPref = function readPref(pref) {
-  switch (pref) {
-    case 'p':
-      pref = 'primary';
-      break;
-    case 'pp':
-      pref = 'primaryPreferred';
-      break;
-    case 's':
-      pref = 'secondary';
-      break;
-    case 'sp':
-      pref = 'secondaryPreferred';
-      break;
-    case 'n':
-      pref = 'nearest';
-      break;
-  }
+const readPrefLookup = {
+  p: 'primary',
+  pp: 'primaryPreferred',
+  s: 'secondary',
+  sp: 'secondaryPreferred',
+  n: 'nearest'
+};
 
-  return pref;
+exports.readPref = function readPref(pref) {
+  return readPrefLookup[pref] || pref;
 };
 
 
@@ -233,27 +238,17 @@ exports.readPref = function readPref(pref) {
  *
  * @param {String|Object} concern
  */
+const readConcernLookup = {
+  l: 'local',
+  a: 'available',
+  m: 'majority',
+  lz: 'linearizable',
+  s: 'snapshot'
+};
 
 exports.readConcern = function readConcern(concern) {
-  if ('string' === typeof concern) {
-    switch (concern) {
-      case 'l':
-        concern = 'local';
-        break;
-      case 'a':
-        concern = 'available';
-        break;
-      case 'm':
-        concern = 'majority';
-        break;
-      case 'lz':
-        concern = 'linearizable';
-        break;
-      case 's':
-        concern = 'snapshot';
-        break;
-    }
-    concern = { level: concern };
+  if (typeof concern === 'string') {
+    concern = { level: readConcernLookup[concern] || concern };
   }
   return concern;
 };
@@ -263,7 +258,7 @@ exports.readConcern = function readConcern(concern) {
  */
 
 const _toString = Object.prototype.toString;
-exports.toString = function(arg) {
+exports.toString = function toString(arg) {
   return _toString.call(arg);
 };
 
@@ -274,8 +269,8 @@ exports.toString = function(arg) {
  * @return {Boolean}
  */
 
-const isObject = exports.isObject = function(arg) {
-  return '[object Object]' == exports.toString(arg);
+const isObject = exports.isObject = function isObject(arg) {
+  return _toString.call(arg) === '[object Object]';
 };
 
 /**
@@ -291,7 +286,7 @@ exports.keys = Object.keys;
  * Based on https://developer.mozilla.org/en-US/docs/JavaScript/Reference/Global_Objects/Object/create
  */
 
-exports.create = 'function' == typeof Object.create
+exports.create = typeof Object.create === 'function'
   ? Object.create
   : create;
 
@@ -309,7 +304,7 @@ function create(proto) {
  * inheritance
  */
 
-exports.inherits = function(ctor, superCtor) {
+exports.inherits = function inherits(ctor, superCtor) {
   ctor.prototype = exports.create(superCtor.prototype);
   ctor.prototype.constructor = ctor;
 };
@@ -319,7 +314,7 @@ exports.inherits = function(ctor, superCtor) {
  * compat with node 0.10 which behaves differently than previous versions
  */
 
-const soon = exports.soon = 'function' == typeof setImmediate
+const soon = exports.soon = typeof setImmediate === 'function'
   ? setImmediate
   : process.nextTick;
 
@@ -330,6 +325,6 @@ const soon = exports.soon = 'function' == typeof setImmediate
  * @return {Boolean}
  */
 
-exports.isArgumentsObject = function(v) {
-  return Object.prototype.toString.call(v) === '[object Arguments]';
+exports.isArgumentsObject = function isArgumentsObject(v) {
+  return _toString.call(v) === '[object Arguments]';
 };