--- a/package.json
+++ b/package.json
@@ -19,6 +19,7 @@
     "debug": "4.x"
   },
   "devDependencies": {
+    "benchmark": "^2.1.4",
     "eslint": "8.x",
     "eslint-plugin-mocha-no-only": "1.1.1",
     "mocha": "9.x",