I also added benchmarks:
before:
```
node clone
clone Binary x 6,092,545 ops/sec ±0.35% (95 runs sampled)
clone Buffer x 1,685,261 ops/sec ±0.80% (89 runs sampled)
clone Date x 3,855,007 ops/sec ±0.21% (95 runs sampled)
clone ObjectId x 13,318,948 ops/sec ±0.59% (95 runs sampled)
clone document x 20,405 ops/sec ±1.05% (92 runs sampled)
clone RegExp x 3,101,792 ops/sec ±0.93% (90 runs sampled)
clone simple array x 46,953 ops/sec ±1.51% (85 runs sampled)
clone big array x 50.22 ops/sec ±1.06% (65 runs sampled)
clone complex array x 174 ops/sec ±2.24% (78 runs sampled)
```

after:
```
node clone
clone Binary x 6,197,530 ops/sec ±0.27% (95 runs sampled)
clone Buffer x 1,846,515 ops/sec ±1.56% (80 runs sampled)
clone Date x 9,323,686 ops/sec ±0.33% (90 runs sampled)
clone ObjectId x 12,511,455 ops/sec ±0.71% (94 runs sampled)
clone document x 92,521 ops/sec ±0.32% (92 runs sampled)
clone RegExp x 8,146,014 ops/sec ±0.34% (93 runs sampled)
clone simple array x 518,904 ops/sec ±0.21% (95 runs sampled)
clone big array x 481 ops/sec ±0.50% (88 runs sampled)
clone complex array x 883 ops/sec ±0.14% (95 runs sampled)
```

It is imho better balanced performance wise. Arrays and documents are significantly faster, which should be the usual business case.